Changes conneted to moved useAuthContext file.
The change in the location of the useAuthContext file also forced a change in the import paths.
