+++ frontend/src/main.tsx
@@ -1,5 +1,6 @@
 import './i18n'
 import { ColorModeProvider } from '@/components/ui/color-mode'
+import { AuthProvider } from '@/contexts/useAuthContext'
 import { ChakraProvider } from '@chakra-ui/react'
 import { MutationCache, QueryCache, QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { RouterProvider, createRouter } from '@tanstack/react-router'
@@ -7,7 +8,6 @@ import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
 import { ApiError, OpenAPI } from './client'
 import AnalyticsConsent from './components/commonUI/AnalyticsConsent'
-import { AuthProvider } from './hooks/useAuthContext'
 import { routeTree } from './routeTree.gen'
 import { system } from './theme'
